requirements.txt
typer~=0.15.2
docker~=7.1.0
fastapi~=0.115.11
GitPython~=3.1.44
aiofiles~=24.1.0
uvicorn~=0.34.0
alembic~=1.15.2